+++ balign/RunAligner.h
@@ -39,7 +39,7 @@ typedef struct {
 void RunAligner(RGBinary*, char*, char*, int, int, int, int, int, int, int, int, int, int, int, int, int, char*, char*, char*, int, int*, int*);
 void RunDynamicProgramming(FILE*, RGBinary*, char*, int, int, int, int, int, int, int, int, int, int, int, int, int, char*, FILE*, FILE*, int, int*, int*);
 void *RunDynamicProgrammingThread(void *);
-void RunDynamicProgrammingThreadHelper(RGBinary*, uint8_t, uint32_t, int8_t, char*, int, int, int, ScoringMatrix*, AlignEntry*);
+void RunDynamicProgrammingThreadHelper(RGBinary*, uint32_t, uint32_t, int8_t, char*, int, int, int, ScoringMatrix*, AlignEntry*);
 void GetSequenceFromReferenceGenome(RGBinary*, int, int, char, int, char*, int, int*, int*);
 void GetReverseComplimentAnyCase(char*, char*, int);
 #endif
